+++ apps/jumentix-website/.storybook/main.ts
@@ -6,11 +6,15 @@ const config: StorybookConfig = {
     disableTelemetry: true,
     enableCrashReports: false,
   },
-  stories: ['../components/**/*.(stories|story).@(js|jsx|ts|tsx)'],
-  addons: ['@storybook/addon-themes'],
+  stories: ['../components/**/*.stories.@(js|jsx|ts|tsx|mdx)', '../components/**/*.story.@(js|jsx|ts|tsx)'],
+  addons: ['@storybook/addon-docs', '@storybook/addon-a11y', '@storybook/addon-themes'],
   framework: {
     name: '@storybook/nextjs',
     options: {},
+  },
+  staticDirs: ['../public'],
+  docs: {
+    defaultName: 'Documentation',
   },
 };
 export default config;